[qca-nss-gmac] 64 bits ethtool stats support

+++ ipq806x/exports/nss_gmac_api_if.h
@@ -45,71 +45,71 @@
  * The NA per-GMAC statistics statistics structure.
  */
 struct nss_gmac_stats {
-	uint32_t rx_bytes;		/**< Number of RX bytes */
-	uint32_t rx_packets;		/**< Number of RX packets */
-	uint32_t rx_errors;		/**< Number of RX errors */
-	uint32_t rx_receive_errors;	/**< Number of RX receive errors */
-	uint32_t rx_overflow_errors;	/**< Number of RX overflow errors */
-	uint32_t rx_descriptor_errors;	/**< Number of RX descriptor errors */
-	uint32_t rx_watchdog_timeout_errors;
+	uint64_t rx_bytes;		/**< Number of RX bytes */
+	uint64_t rx_packets;		/**< Number of RX packets */
+	uint64_t rx_errors;		/**< Number of RX errors */
+	uint64_t rx_receive_errors;	/**< Number of RX receive errors */
+	uint64_t rx_overflow_errors;	/**< Number of RX overflow errors */
+	uint64_t rx_descriptor_errors;	/**< Number of RX descriptor errors */
+	uint64_t rx_watchdog_timeout_errors;
 					/**< Number of RX watchdog timeout errors */
-	uint32_t rx_crc_errors;		/**< Number of RX CRC errors */
-	uint32_t rx_late_collision_errors;
+	uint64_t rx_crc_errors;		/**< Number of RX CRC errors */
+	uint64_t rx_late_collision_errors;
 					/**< Number of RX late collision errors */
-	uint32_t rx_dribble_bit_errors;	/**< Number of RX dribble bit errors */
-	uint32_t rx_length_errors;	/**< Number of RX length errors */
-	uint32_t rx_ip_header_errors;	/**< Number of RX IP header errors */
-	uint32_t rx_ip_payload_errors;	/**< Number of RX IP payload errors */
-	uint32_t rx_no_buffer_errors;	/**< Number of RX no-buffer errors */
-	uint32_t rx_transport_csum_bypassed;
+	uint64_t rx_dribble_bit_errors;	/**< Number of RX dribble bit errors */
+	uint64_t rx_length_errors;	/**< Number of RX length errors */
+	uint64_t rx_ip_header_errors;	/**< Number of RX IP header errors */
+	uint64_t rx_ip_payload_errors;	/**< Number of RX IP payload errors */
+	uint64_t rx_no_buffer_errors;	/**< Number of RX no-buffer errors */
+	uint64_t rx_transport_csum_bypassed;
 					/**< Number of RX packets where the transport checksum was bypassed */
-	uint32_t tx_bytes;		/**< Number of TX bytes */
-	uint32_t tx_packets;		/**< Number of TX packets */
-	uint32_t tx_collisions;		/**< Number of TX collisions */
-	uint32_t tx_errors;		/**< Number of TX errors */
-	uint32_t tx_jabber_timeout_errors;
+	uint64_t tx_bytes;		/**< Number of TX bytes */
+	uint64_t tx_packets;		/**< Number of TX packets */
+	uint64_t tx_collisions;		/**< Number of TX collisions */
+	uint64_t tx_errors;		/**< Number of TX errors */
+	uint64_t tx_jabber_timeout_errors;
 					/**< Number of TX jabber timeout errors */
-	uint32_t tx_frame_flushed_errors;
+	uint64_t tx_frame_flushed_errors;
 					/**< Number of TX frame flushed errors */
-	uint32_t tx_loss_of_carrier_errors;
+	uint64_t tx_loss_of_carrier_errors;
 					/**< Number of TX loss of carrier errors */
-	uint32_t tx_no_carrier_errors;	/**< Number of TX no carrier errors */
-	uint32_t tx_late_collision_errors;
+	uint64_t tx_no_carrier_errors;	/**< Number of TX no carrier errors */
+	uint64_t tx_late_collision_errors;
 					/**< Number of TX late collision errors */
-	uint32_t tx_excessive_collision_errors;
+	uint64_t tx_excessive_collision_errors;
 					/**< Number of TX excessive collision errors */
-	uint32_t tx_excessive_deferral_errors;
+	uint64_t tx_excessive_deferral_errors;
 					/**< Number of TX excessive deferral errors */
-	uint32_t tx_underflow_errors;	/**< Number of TX underflow errors */
-	uint32_t tx_ip_header_errors;	/**< Number of TX IP header errors */
-	uint32_t tx_ip_payload_errors;	/**< Number of TX IP payload errors */
-	uint32_t tx_dropped;		/**< Number of TX dropped packets */
-	uint32_t hw_errs[10];		/**< GMAC DMA error counters */
-	uint32_t rx_missed;		/**< Number of RX packets missed by the DMA */
-	uint32_t fifo_overflows;	/**< Number of RX FIFO overflows signalled by the DMA */
-	uint32_t rx_scatter_errors;	/**< Number of scattered frames received by the DMA */
-	uint32_t tx_ts_create_errors;	/**< Number of tx timestamp creation errors */
-	uint32_t gmac_total_ticks;	/**< Total clock ticks spend inside the GMAC */
-	uint32_t gmac_worst_case_ticks;	/**< Worst case iteration of the GMAC in ticks */
-	uint32_t gmac_iterations;	/**< Number of iterations around the GMAC */
-	uint32_t tx_pause_frames;	/**< Number of pause frames sent by the GMAC */
-	uint32_t rx_octets_g;		/* Number of good octets received */
-	uint32_t rx_ucast_frames;	/* Number of Unicast frames received */
-	uint32_t rx_bcast_frames;	/* Number of Bcast frames received */
-	uint32_t rx_mcast_frames;	/* Number of Mcast frames received */
-	uint32_t rx_undersize;		/* Number of RX undersize frames */
-	uint32_t rx_oversize;		/* Number of RX oversize frames */
-	uint32_t rx_jabber;		/* Number of jabber frames */
-	uint32_t rx_octets_gb;		/* Number of good/bad octets */
-	uint32_t rx_frag_frames_g;	/* Number of good ipv4 frag frames */
-	uint32_t tx_octets_g;		/* Number of good octets sent */
-	uint32_t tx_ucast_frames;	/* Number of Unicast frames sent*/
-	uint32_t tx_bcast_frames;	/* Number of Broadcast frames sent */
-	uint32_t tx_mcast_frames;	/* Number of Multicast frames sent */
-	uint32_t tx_deferred;		/* Number of Deferred frames sent */
-	uint32_t tx_single_col;		/* Number of single collisions */
-	uint32_t tx_multiple_col;	/* Number of multiple collisions */
-	uint32_t tx_octets_gb;		/* Number of good/bad octets sent*/
+	uint64_t tx_underflow_errors;	/**< Number of TX underflow errors */
+	uint64_t tx_ip_header_errors;	/**< Number of TX IP header errors */
+	uint64_t tx_ip_payload_errors;	/**< Number of TX IP payload errors */
+	uint64_t tx_dropped;		/**< Number of TX dropped packets */
+	uint64_t hw_errs[10];		/**< GMAC DMA error counters */
+	uint64_t rx_missed;		/**< Number of RX packets missed by the DMA */
+	uint64_t fifo_overflows;	/**< Number of RX FIFO overflows signalled by the DMA */
+	uint64_t rx_scatter_errors;	/**< Number of scattered frames received by the DMA */
+	uint64_t tx_ts_create_errors;	/**< Number of tx timestamp creation errors */
+	uint64_t gmac_total_ticks;	/**< Total clock ticks spend inside the GMAC */
+	uint64_t gmac_worst_case_ticks;	/**< Worst case iteration of the GMAC in ticks */
+	uint64_t gmac_iterations;	/**< Number of iterations around the GMAC */
+	uint64_t tx_pause_frames;	/**< Number of pause frames sent by the GMAC */
+	uint64_t rx_octets_g;		/* Number of good octets received */
+	uint64_t rx_ucast_frames;	/* Number of Unicast frames received */
+	uint64_t rx_bcast_frames;	/* Number of Bcast frames received */
+	uint64_t rx_mcast_frames;	/* Number of Mcast frames received */
+	uint64_t rx_undersize;		/* Number of RX undersize frames */
+	uint64_t rx_oversize;		/* Number of RX oversize frames */
+	uint64_t rx_jabber;		/* Number of jabber frames */
+	uint64_t rx_octets_gb;		/* Number of good/bad octets */
+	uint64_t rx_frag_frames_g;	/* Number of good ipv4 frag frames */
+	uint64_t tx_octets_g;		/* Number of good octets sent */
+	uint64_t tx_ucast_frames;	/* Number of Unicast frames sent*/
+	uint64_t tx_bcast_frames;	/* Number of Broadcast frames sent */
+	uint64_t tx_mcast_frames;	/* Number of Multicast frames sent */
+	uint64_t tx_deferred;		/* Number of Deferred frames sent */
+	uint64_t tx_single_col;		/* Number of single collisions */
+	uint64_t tx_multiple_col;	/* Number of multiple collisions */
+	uint64_t tx_octets_gb;		/* Number of good/bad octets sent*/
 };
 
 /*
